zmiany w angle_chech (dalej nie działa)

+++ src/Ant_worker.java
@@ -24,7 +24,7 @@ public class Ant_worker extends Ant{
         x = Math.round(x * (4 / distance));                                 //scales x and y movement to move aproximetly 2 tiles
         y = Math.round(y * (4 / distance));
 
-        this.position.change_coordinates((short)x, (short)y);               //changes its position
+        this.position.change_coordinates(x, y);               //changes its position
 
         return false;
     }
